Move the new classes into sub packages
The main package is getting big.
