+++ srcs/juloo.keyboard2/EditorConfig.java
@@ -5,7 +5,7 @@ import android.os.Build.VERSION;
 import android.text.InputType;
 import android.text.TextUtils;
 import android.view.inputmethod.EditorInfo;
-import juloo.keyboard2.CandidatesView;
+import juloo.keyboard2.suggestions.CandidatesView;
 
 public final class EditorConfig
 {
